requirements.txt
Django==2.2
django-cors-headers==3.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2-binary==2.8.4
python-dotenv==0.10.3